# api-spec.yaml
openapi: 3.0.0
info:
  title: Quiz Game API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server paths

paths:
  /games:
    post:
      summary: Create a new game session
      responses:
        '201':
          description: Game session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'

  /games/{gameId}/players:
    post:
      summary: Join a game session
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Player information
        required: true
        content:
          application/json:
            schema:
              properties:
                code:
                  type: string
                name:
                  type: string
      responses:
        '200':
          description: Player joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'

  /games/{gameId}/start:
    post:
      summary: Start the game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'

  /games/{gameId}/current-question:
    get:
      summary: Get the current question
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'

  /games/{gameId}/answers:
    post:
      summary: Submit an answer
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Answer submission
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Answer'
      responses:
        '200':
          description: Answer received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'

  /games/{gameId}/leaderboard:
    get:
      summary: Get the leaderboard
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'

components:
  schemas:
    GameSession:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [waiting, active, finished]
        currentQuestionId:
          type: string
        startTime:
          type: string
          format: date-time

    Player:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        score:
          type: integer
        joinedAt:
          type: string
          format: date-time
        gameId:
          type: string

    Question:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: string
        timeLimit:
          type: integer
          description: Time limit in seconds

    Answer:
      type: object
      properties:
        playerId:
          type: string
        questionId:
          type: string
        selectedOption:
          type: integer
        responseTime:
          type: integer
          description: Response time in milliseconds

    AnswerResponse:
      type: object
      properties:
        correct:
          type: boolean
        correctOption:
          type: integer
        scoreAwarded:
          type: integer

    LeaderboardEntry:
      type: object
      properties:
        playerId:
          type: string
        name:
          type: string
        score:
          type: integer